Ensure examples in pydocs run successfuly
The motivation for this is that these examples act as documentation, and
so it's nice to be able to validate they behave as expected. To avoid
regressions you could run these as part of `pytest` via the
`--doctest-modules` flag.

Fixing these mostly involved fixing up the formatting of some output, or
importing some modules that aren't already in the module's scope. I also
had to move some comments since if these are in the output line
`doctest` expects that literal string in the output.

One test was skipped since it was for a function returning a random
value.

There were also some examples where the input didn't give the expected
outputs, in theses cases I aimed to fix up the input (there wasn't a
particular reason to do this over changing the output). These were:

* In `xocto.localetime` set a `datetime`'s `minute` attribute since `21`
  is closer to `15` than `30`
* In `xocto.ranges` use `decimal.Decimal` objects as input where
  expected instead of floats
* In `xocto.ranges` fix an invalid range since `Range(None, 2)` raises:

    ValueError: Range with unbounded start must be left-exclusive

Note, I used strings as arguments to `Decimal` to avoid having to deal
with rounding issues, e.g.

    >>> from xocto import numbers
    >>> from decimal import Decimal
    >>> Decimal(12.35)
    Decimal('12.3499999999999996447286321199499070644378662109375')
    >>> numbers.round_decimal_places(Decimal(12.35), 1)
    Decimal('12.3')
    >>> numbers.round_decimal_places(Decimal('12.35'), 1)
    Decimal('12.4')

This also raised an inconsistency between the pydocs and the behaviour
of `decimal.ROUND_HALF_UP`[1], which:

> Round to nearest with ties going away from zero.

So I updated these docs too.

[1] https://docs.python.org/3.8/library/decimal.html#decimal.ROUND_HALF_UP